==> RQ2-Data_Extraction/1_Group_the_Families_according_to_AIHub_info-WithAuthor.ipynb <==
In this script, some raw classification for families will be done. 

import os, json
from pprint import pprint
import pandas as pd

## See how many models have multiple versions: 

def see_how_many_models_have_multiple_versions():
    raw_file_path = "J:/EMSE2020_MLPackage/Data/TFHub/2020-03-10/TFHubModule_Info.json"
    with open(raw_file_path, "r") as rfp:
        modules = json.load(rfp)
    
    specialFormatModules = []
    
    module_names = {}
    for module in modules:
        module_name = module["name"]
        version_num = module["version"]
        author = None
        for l in module["info_NormalDeployment"][-1][42]:
            if l[0] == "publisher":
                author = l[1][0][0]
                break
        module_name = "{}@@{}".format(module_name, author)
        
#                 print(l[1][1])
        
        moduleFormat = module["info_NormalDeployment"][-1][-11]
        ## 我们只关注默认格式是这两种的package，同时，我们也只关注它们的这两种格式的模型。
        ## 举例，有的package默认格式就是TFLite，TFJS之类，没有普通格式hub和tf2sm，那我们不管这些package。 https://tfhub.dev/tensorflow/coral-model/mobilenet_v2_1.0_224_quantized/1/default/1
        ## 同时，有一个package有tf2sm格式，同时还有tflite和tfjs格式，那我们只关注tf2sm格式，不关注另外的格式。https://tfhub.dev/google/imagenet/mobilenet_v2_075_96/feature_vector/4
        ## 不知道这样可不可以。
        if moduleFormat not in ("saved_model_2", "hub", "saved_model"):
            specialFormatModules.append(module_name)
            continue
            
        if module_name not in module_names:
            module_names[module_name] = [version_num]
        else:
            module_names[module_name].append(version_num)

    keys = list(module_names.keys())
    keys.sort()
    print("there are {} modules in TFHub\n\n\n".format(len(keys)))
    versionCount = 0
    for name in keys:
        v_list = module_names[name]
        versionCount += len(v_list)
        print(name, v_list)
#         if len(v_list) > 1:
#             print(name, v_list)
    print("There are in total {} versions in TFHub".format(versionCount))
#     print("Those modules are not in general formats: ")
#     pprint(set(specialFormatModules))
    
    
see_how_many_models_have_multiple_versions()

## Generate a CSV table. 

def generate_a_csv_table():
    raw_file_path = "J:/ModelStoreData/AIHub/2020-03-10/TFHubModule_Info.json"
    with open(raw_file_path, "r") as rfp:
        modules = json.load(rfp)
    with open("rst.csv", "w") as rst:
        rst.write("{},{},{},{},{},{},{}\n".format("name", "version", "dataset", "fmt", "module_type", "network_architecture", "author"))
        for module in modules:
            name = module["name"]
            version = module["version"]
            author = None
            for l in module["info_NormalDeployment"][-1][42]:
                if l[0] == "publisher":
                    author = l[1][0][0]
                    break
            name = "{}@@{}".format(name, author)
            
            ## 排除掉一些特异格式的模型。
            moduleFormat = module["info_NormalDeployment"][-1][-11]
            if moduleFormat not in ("saved_model_2", "hub", "saved_model"):
                continue
            
            labels = module["info_NormalDeployment"][1][19]
#             print(labels)
            dataset = None
            fmt = moduleFormat #None
            module_type = None
            network_architecture = None
            for _ in labels:
                label = _.replace(",", "|")
                if "dataset" in label and dataset == None:
                    dataset = label.split(":")[1]
#                 elif "format" in label:
#                     print(name, version)
#                     if "hub" in label:
#                         fmt = "hub"
#                         continue
#                     else: 
#                         fmt = label.split(":")[1]
                elif "module-type" in label and module_type == None:
                    module_type = label.split(":")[1]
                elif "network-architecture" in label and network_architecture == None:
                    network_architecture = label.split(":")[1]
            # print("{} {} {} {} {} {}\n".format(name, version, dataset, fmt, module_type, network_architecture))
            rst.write("{},{},{},{},{},{},{}\n".format(name, version, dataset, fmt, module_type, network_architecture, author))
generate_a_csv_table()

## Look into the CSV file, some Database operations

It seems that there are around `62%` modules have multiple versions.

def how_many_models_have_multiple_versions():
    table = pd.read_csv("rst.csv")
    table.head()
    ##---------------
    df_gp = table.groupby("name", as_index=False)
    new_df = df_gp.count().reset_index()
    # new_df.head()
    print(
        "\nThere are in total {} modules.".format(new_df.shape[0])
    )
    ## --------------------
    new_df.drop(columns=["index", "dataset", "fmt", "module_type", "network_architecture", "author"], inplace=True)
    new_df.rename(columns={"version": "versions"}, inplace=True)
    print(new_df.head())
    new_df.to_csv("how_many_models_have_multiple_versions.csv", index=False)
    print(
        "\nThere are in total {} modules have multiple versions.".format(new_df[new_df["versions"] > 1].shape[0])
    )
    return

how_many_models_have_multiple_versions()



def how_many_models_have_multiple_format():
    table = pd.read_csv("rst.csv")
    table.head()
    ## --------------------
    df_gp = table.groupby(["name", "fmt"], as_index=False)
    new_df = df_gp.count().reset_index()
    new_df.drop(columns=["index", "dataset", "module_type", "network_architecture", "author"], inplace=True)
    # print(new_df[new_df["name"].str.contains("classification")])
    ### --------------------
    df_gp = new_df.groupby(["name"], as_index=False)
    new_df = df_gp.count().reset_index()
    new_df.drop(columns=["index", "version"], inplace=True)
    ### ------------------
    new_df.rename(columns={"fmt": "fmts"}, inplace=True)
    print(new_df.head())
    new_df.to_csv("how_many_models_have_multiple_format.csv", index=False)
    print(
        "\nThere are in total {} modules have multiple formats.".format(new_df[new_df["fmts"] > 1].shape[0])
    )
    return

how_many_models_have_multiple_format()

def how_many_models_have_multiple_format1():
    table = pd.read_csv("rst.csv")
    table.head()
    ## --------------------
    df_gp = table.groupby(["name", "fmt"], as_index=False)
    new_df = df_gp.count().reset_index()
    new_df.drop(columns=["index", "dataset", "module_type", "network_architecture", "author"], inplace=True)
    # print(new_df[new_df["name"].str.contains("classification")])
    ### --------------------
    df_gp = new_df.groupby(["name"], as_index=False)
    new_df = df_gp.count().reset_index()
    new_df.drop(columns=["index", "version"], inplace=True)
    new_df["fmt_str"] = df_gp['fmt'].apply(','.join)
    ### ------------------
    new_df.rename(columns={"fmt": "fmts"}, inplace=True)
    print(new_df.head())
    new_df.to_csv("how_many_models_have_multiple_format.csv", index=False)
    print(
        "\nThere are in total {} modules have multiple formats.".format(new_df[new_df["fmts"] > 1].shape[0])
    )
    return

how_many_models_have_multiple_format1()

def module_type_and_network_architecture_of_the_modules():
    table = pd.read_csv("rst.csv")
    table.head()
    ## --------------------
    df_gp = table.groupby(["name", "dataset", "module_type", "network_architecture"], as_index=False)
    new_df = df_gp.count().reset_index()
    new_df.drop(columns=["index", "fmt", "version"], inplace=True)
    ## --------------------
    print(new_df.head())
    new_df.to_csv("module_type_and_network_architecture_of_the_modules.csv", index=False)
    print(
        "\nThere are in total {} modules have other fields.".format(new_df.shape[0])
    )
    return

module_type_and_network_architecture_of_the_modules()

## Merge the CSV files and see some statistics

I don't know why in the merged result, universal-sentence-encoder-lite will have an extra line. 奇哉怪哉。 

def merge_the_csv_files(): 
    table1 = pd.read_csv("how_many_models_have_multiple_versions.csv")
    table2 = pd.read_csv("how_many_models_have_multiple_format.csv")
    table3 = pd.read_csv("module_type_and_network_architecture_of_the_modules.csv")

    _1 = pd.merge(table1, table2, "inner", on="name")
    _2 = pd.merge(_1, table3, "inner", on="name")

    merged_table = _2
    print(merged_table.shape[0])
    merged_table.to_csv("final_info.csv", index=False)
    return merged_table

merged_table = merge_the_csv_files()
merged_table.head()

def sort_merged_table(merged_table):
    merged_table = merged_table[["module_type", "network_architecture", "dataset", "name", "versions", "fmts", "fmt_str", "author"]]
    merged_table.sort_values(["module_type", "network_architecture", "dataset", "name"], ascending=True, inplace=True)
#     merged_table.sort_values(["name"], ascending=True, inplace=True)
    merged_table.to_csv("sorted.csv", index=False)
    return merged_table

sorted_table = sort_merged_table(merged_table)

## 从模型的名字里面把作者拆出来。
def getAuthors(sorted_table):
    sorted_table["_"], sorted_table["author"] = sorted_table["name"].str.split("@@").str
    sorted_table.drop(columns=["_"], inplace=True)
    sorted_table.to_csv("sorted-withAuthor.csv", index=False)
    return
getAuthors(sorted_table)

merged_table["module_type"].value_counts()
