==> RQ2-Data_Extraction/3-Group_family_second_step.ipynb <==
Need to run `1_xxx.ipynb` first. 

## Group descriptions according to families

import os, json, difflib
from pprint import pprint
import pandas as pd

table = pd.read_csv(r"sorted.csv")
def find_info_for_module(table, module, column):
#     try:
    return (table[table["name"] == module][column].tolist()[0])
#     except:
#         print(module)
# find_info_for_module(table, "efficientnet-b3-classification", "module_type")

def family_name_decision(module):
    """
    return the family name of the module
    """
    families = [
        "Wiki-words",
        "albert_" , 
        "bert_"  ,
        "bigbigan-" , 
        "biggan-"  ,
#         "biggan-deep"  ,
        "compare_gan-model"  ,
        "compare_gan-s3gan"  ,
        "efficientnet-b" , 
        "i3d-kinetics"  ,
        "image_augmentation"  ,
        "image_augmentation-nas"  ,
        "imagenet-amoebanet"  ,
        "imagenet-inception_resnet_v2"  ,
        "imagenet-inception_v1"  ,
        "imagenet-inception_v2"  ,
        "imagenet-inception_v3"  ,
        "imagenet-mobilenet_v1"  ,
#         "quantops"  ,
        "imagenet-mobilenet_v2"  ,
        "imagenet-nasnet"  ,
        "imagenet-pnasnet"  ,
        "imagenet-resnet_v1"  ,
        "imagenet-resnet_v2"  ,
        "nnlm"  ,
#         "random-nnlm"  ,
        "tf2-preview-gnews-swivel" ,
        "tf2-preview-inception_v3" ,
        "tf2-preview-mobilenet_v2" ,
#         "tf2-preview-nnlm",
        "universal-sentence-encoder"  ,
#         "universal-sentence-encoder-xling"
       ]
    for family in families:
        if family in module:
            print()
    
    
path = r"module_descriptions"    
path_to = r"family_raw_grouping"
if not os.path.exists(path_to):
    os.makedirs(path_to)
for module in os.listdir(path):
    path_lv1 = os.path.join(path, module)
    whole_file = ""
    version_counter = 0
    for file in reversed(os.listdir(path_lv1)):
        if ".md" not in file:
            continue
        version_counter += 1
        path_lv2 = os.path.join(path_lv1, file)
        with open(path_lv2, "r", encoding="utf-8") as pl:
            content = pl.read()
        whole_file += "\n\n# _{}_\n{}".format(file, content)
    whole_file = "There are `{}` versions in total.".format(version_counter) + whole_file
    task_type = find_info_for_module(table, module, "module_type")
    path_lv1_1 = os.path.join(path_to, task_type)
    if not os.path.exists(path_lv1_1):
        os.makedirs(path_lv1_1)
    with open (os.path.join(path_lv1_1, module+".md"), "w", encoding="utf-8") as ptm:
        ptm.write(whole_file)
#     break

## Manually group the family members. See the folder `family_grouping-final`
