==> RQ2-Data_Extraction/5-GoThroughMDFiles.ipynb <==
In this notebook, we go through all of the files to have some findings. 

import os, shutil

def generateABigFile_seeWhatPackageDontMentionSignature(p):
    counter = 0
    if not os.path.exists(r"J:\EMSE2020_MLPackage\RQ2-Data_Extraction\no_sig_files"): 
        os.makedirs(r"J:\EMSE2020_MLPackage\RQ2-Data_Extraction\no_sig_files")
    for path,dir_list,file_list in os.walk(p):  
        for file_name in file_list:  
            if ".md" not in file_name:
                continue
            path_lv1 = os.path.join(path, file_name)
            no_path = path_lv1.replace("J:\\EMSE2020_MLPackage\\RQ2-Data_Extraction\\family_grouping-final\\", "")
            name = no_path.split("@@")[0]
            # print(name)
            with open(path_lv1, "r", encoding="utf-8") as pl:
                lines = pl.readlines()
            lines = [line.strip() for line in lines]
            new_lines = []
            code_zone = False
            ## 不要代码
            for line in lines:
                if "```" in line and code_zone == False:
                    code_zone = True
                    continue
                elif "```" in line and code_zone == True:
                    code_zone = False
                    continue
                else:
                    if code_zone == True:
                        continue
                    else:
                        new_lines.append(line)
            oneLine = " ".join(new_lines)
            ## 
            if "ignature" not in oneLine:
#                 print(path_lv1)
                counter += 1
                shutil.copy(
                    path_lv1, 
                    r"J:\EMSE2020_MLPackage\RQ2-Data_Extraction\no_sig_files\{}-{}.md".format(counter, file_name)
                ) 

generateABigFile_seeWhatPackageDontMentionSignature(r"J:\EMSE2020_MLPackage\RQ2-Data_Extraction\family_grouping-final")

After running the last cell, we can find 160 files that don't mention "signature". 

Which indicates that more than half of the files mention "signature", but they don't list them as a formalized list.
